alphanet2: use real node names instead of going through reverse proxy
Going through the reverse proxy @ alphanet2.radixdlt.com is very
risky (centralised). We need to start using valid node names asap.

Professional node-runners are expected to setup their own DNS records,
which is why the resolveNodeName should allow those to pass through
as well.

In regards to IPv6, which is not supported for alphanet2, the
resolveNodeName function will not calculate a valid node name at this
point. This will be fixed at a later time.